# api/openapi-go-away-2024.yml
openapi: 3.0.4
info:
  title: Advent of Code (go-away-2024)
  version: 1.0.0
tags:
  - name: AdventOfCode
    description: Task management
paths:
  /task/create:
    post: 
      tags:
        - AdventOfCode
      summary: Create task to solve
      operationId: postTask
      parameters:
        - name: year
          in: query
          description: Year of AoC challenge
          required: true
          schema: 
            type: integer
            format: int32
            example: 2024
        - name: day
          in: query
          description: Puzzle day
          required: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: part
          in: query
          description: Puzzle part (1 or 2)
          required: true
          schema:
            type: integer
            format: int32
            example: 1
      requestBody:
        description: Data for task solving
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                filename: 
                  type: string
                  format: binary
      responses:
        "200":
          description: Successful operation 
          content:
            applicaton/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: Unexpected error
          content:
            applicaton/json:
              schema:
                $ref: "#/components/schemas/Error"
  /task/{id}:
    get:
      tags:
        - AdventOfCode
      summary: Get task status
      operationId: getTask
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
            pattern: ^\d+$
            example: 666
      responses:
        "200":
          description: Successful operation 
          content:
            applicaton/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: Unexpected error
          content:
            applicaton/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Task:
      required:
        - id 
        - status
        - message
        - createdAt
      type: object
      properties:
        id:
          description: Task ID
          type: string
          pattern: ^\d+$
          example: 666
        status:
          description: Task solution status
          type: string
          enum:
            - CREATED 
            - STARTED
            - COMPLITED
            - ERROR 
          example: SOLVED
        result:
          description: Task solution result
          type: string
          example: 666
        message: 
          description: Task solution message
          type: string
          example: Day 1 Part 1 from AoC 2024 successfully solved with the following result
        createdAt:
          description: Task creation date 
          type: string
          format: date-time 
          example: 2025-11-22T00:00:00.000+03:00
        startedAt:
          description: Start date of the task solving
          type: string
          format: date-time 
          example: 2025-11-22T00:01:00.000+03:00
        completedAt:
          description: Completion date of the task solving
          type: string
          format: date-time 
          example: 2025-11-22T00:02:00.000+03:00
    Error:
      required:
        - code 
        - message 
      type: object
      properties:
        code: 
          description: Error code 
          type: integer
          format: int32
          example: 500
        message:
          description: Error message
          type: string
          example: Puzzle is not supported