# internal/server/aoc.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package server

import (
	"fmt"
	"net/url"
	"time"

	"github.com/gofiber/fiber/v2"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for TaskStatus.
const (
	COMPLITED TaskStatus = "COMPLITED"
	CREATED   TaskStatus = "CREATED"
	ERROR     TaskStatus = "ERROR"
	STARTED   TaskStatus = "STARTED"
)

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// Task defines model for Task.
type Task struct {
	// CompletedAt Completion date of the task solving
	CompletedAt *time.Time `json:"completedAt,omitempty"`

	// CreatedAt Task creation date
	CreatedAt time.Time `json:"createdAt"`

	// Id Task ID
	Id string `json:"id"`

	// Message Task solution message
	Message string `json:"message"`

	// Result Task solution result
	Result *string `json:"result,omitempty"`

	// StartedAt Start date of the task solving
	StartedAt *time.Time `json:"startedAt,omitempty"`

	// Status Task solution status
	Status TaskStatus `json:"status"`
}

// TaskStatus Task solution status
type TaskStatus string

// PostTaskMultipartBody defines parameters for PostTask.
type PostTaskMultipartBody struct {
	Filename *openapi_types.File `json:"filename,omitempty"`
}

// PostTaskParams defines parameters for PostTask.
type PostTaskParams struct {
	// Year Year of AoC challenge
	Year int32 `form:"year" json:"year"`

	// Day Puzzle day
	Day int32 `form:"day" json:"day"`

	// Part Puzzle part (1 or 2)
	Part int32 `form:"part" json:"part"`
}

// PostTaskMultipartRequestBody defines body for PostTask for multipart/form-data ContentType.
type PostTaskMultipartRequestBody PostTaskMultipartBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create task to solve
	// (POST /task/create)
	PostTask(c *fiber.Ctx, params PostTaskParams) error
	// Get task status
	// (GET /task/{id})
	GetTask(c *fiber.Ctx, id string) error
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

type MiddlewareFunc fiber.Handler

// PostTask operation middleware
func (siw *ServerInterfaceWrapper) PostTask(c *fiber.Ctx) error {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params PostTaskParams

	var query url.Values
	query, err = url.ParseQuery(string(c.Request().URI().QueryString()))
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for query string: %w", err).Error())
	}

	// ------------- Required query parameter "year" -------------

	if paramValue := c.Query("year"); paramValue != "" {

	} else {
		err = fmt.Errorf("Query argument year is required, but not found")
		c.Status(fiber.StatusBadRequest).JSON(err)
		return err
	}

	err = runtime.BindQueryParameter("form", true, true, "year", query, &params.Year)
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter year: %w", err).Error())
	}

	// ------------- Required query parameter "day" -------------

	if paramValue := c.Query("day"); paramValue != "" {

	} else {
		err = fmt.Errorf("Query argument day is required, but not found")
		c.Status(fiber.StatusBadRequest).JSON(err)
		return err
	}

	err = runtime.BindQueryParameter("form", true, true, "day", query, &params.Day)
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter day: %w", err).Error())
	}

	// ------------- Required query parameter "part" -------------

	if paramValue := c.Query("part"); paramValue != "" {

	} else {
		err = fmt.Errorf("Query argument part is required, but not found")
		c.Status(fiber.StatusBadRequest).JSON(err)
		return err
	}

	err = runtime.BindQueryParameter("form", true, true, "part", query, &params.Part)
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter part: %w", err).Error())
	}

	return siw.Handler.PostTask(c, params)
}

// GetTask operation middleware
func (siw *ServerInterfaceWrapper) GetTask(c *fiber.Ctx) error {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Params("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter id: %w", err).Error())
	}

	return siw.Handler.GetTask(c, id)
}

// FiberServerOptions provides options for the Fiber server.
type FiberServerOptions struct {
	BaseURL     string
	Middlewares []MiddlewareFunc
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router fiber.Router, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, FiberServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router fiber.Router, si ServerInterface, options FiberServerOptions) {
	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	for _, m := range options.Middlewares {
		router.Use(fiber.Handler(m))
	}

	router.Post(options.BaseURL+"/task/create", wrapper.PostTask)

	router.Get(options.BaseURL+"/task/:id", wrapper.GetTask)

}